# Conflicts:
#	src/main/java/com/shophelperjavasb/shophelperjavasb/shippers/model/Shipper.java
#	src/main/java/com/shophelperjavasb/shophelperjavasb/shippers/repositories/ShippersRepository.java
#	src/main/java/com/shophelperjavasb/shophelperjavasb/users/services/impl/UsersServiceImpl.java

+++ src/main/java/com/shophelperjavasb/shophelperjavasb/products/repositories/ProductsRepository.java
@@ -6,9 +6,9 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface ProductsRepository extends JpaRepository<Product, Integer> {
+public interface ProductsRepository extends JpaRepository<Product, Long> {
     @Override
-    Optional<Product> findById(Integer id);
+    Optional<Product> findById(Long id);
 
     List<Product> findByName(String productName);
 }
