+++ src/main/java/com/shophelperjavasb/shophelperjavasb/shippers/repositories/ShippersRepository.java
@@ -3,6 +3,5 @@ package com.shophelperjavasb.shophelperjavasb.shippers.repositories;
 import com.shophelperjavasb.shophelperjavasb.shippers.model.Shipper;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ShippersRepository extends JpaRepository<Shipper, Long> {
-
+public interface ShippersRepository extends JpaRepository<Shipper, Integer> {
 }
